+++ crates/forge/tests/integration/declare.rs
@@ -7,54 +7,6 @@ use indoc::indoc;
 use std::path::Path;
 
 #[test]
-fn simple() {
-    let test = test_case!(indoc!(
-        r#"#[test]
-        fn test_two_and_two() {
-            assert(2 == 2, '2 == 2');
-        }
-    "#
-    ));
-
-    let result = run(
-        &test.path().unwrap(),
-        &test.path().unwrap().join("src/lib.cairo"),
-        &Some(test.linked_libraries()),
-        &Default::default(),
-        &corelib_path(),
-        &test.contracts(&corelib_path()).unwrap(),
-        &Utf8PathBuf::from_path_buf(predeployed_contracts().to_path_buf()).unwrap(),
-    )
-    .unwrap();
-
-    assert_passed!(result);
-}
-
-#[test]
-fn failing() {
-    let test = test_case!(indoc!(
-        r#"#[test]
-        fn test_two_and_three() {
-            assert(2 == 3, '2 == 3');
-        }
-    "#
-    ));
-
-    let result = run(
-        &test.path().unwrap(),
-        &test.path().unwrap().join("src/lib.cairo"),
-        &Some(test.linked_libraries()),
-        &Default::default(),
-        &corelib_path(),
-        &test.contracts(&corelib_path()).unwrap(),
-        &Utf8PathBuf::from_path_buf(predeployed_contracts().to_path_buf()).unwrap(),
-    )
-    .unwrap();
-
-    assert_failed!(result);
-}
-
-#[test]
 fn simple_declare() {
     let test = test_case!(
         indoc!(
@@ -62,6 +14,7 @@ fn simple_declare() {
         use result::ResultTrait;
         use traits::Into;
         use starknet::ClassHashIntoFelt252;
+        use cheatcodes::declare;
 
         #[test]
         fn test_declare_simple() {
@@ -121,6 +74,7 @@ fn multiple_declare() {
         use result::ResultTrait;
         use traits::Into;
         use starknet::ClassHashIntoFelt252;
+        use cheatcodes::declare;
 
         #[test]
         fn multiple_contracts() {
@@ -224,6 +178,7 @@ fn simple_declare_from_contract_code() {
         use result::ResultTrait;
         use traits::Into;
         use starknet::ClassHashIntoFelt252;
+        use cheatcodes::declare;
 
         #[test]
         fn test_declare_simple() {
@@ -257,6 +212,7 @@ fn declare_unknown() {
         use result::ResultTrait;
         use traits::Into;
         use starknet::ClassHashIntoFelt252;
+        use cheatcodes::declare;
 
         #[test]
         fn test_declare_simple() {
