+++ crates/forge/tests/integration/dispatchers.rs
@@ -19,7 +19,7 @@ fn simple_call_and_invoke() {
         use traits::TryInto;
         use starknet::ContractAddress;
         use starknet::Felt252TryIntoContractAddress;
-        use cheatcodes::PreparedContract;
+        use cheatcodes::{ declare, PreparedContract, deploy };
             
         #[starknet::interface]
         trait IHelloStarknet<TContractState> {
@@ -78,7 +78,7 @@ fn advanced_types() {
         use traits::TryInto;
         use starknet::ContractAddress;
         use starknet::Felt252TryIntoContractAddress;
-        use cheatcodes::PreparedContract;
+        use cheatcodes::{ declare, PreparedContract, deploy, start_prank };
             
 
         #[starknet::interface]
@@ -162,7 +162,7 @@ fn handling_errors() {
         use traits::TryInto;
         use starknet::ContractAddress;
         use starknet::Felt252TryIntoContractAddress;
-        use cheatcodes::PreparedContract;
+        use cheatcodes::{ declare, PreparedContract, deploy };
             
         #[starknet::interface]
         trait IHelloStarknet<TContractState> {
@@ -239,6 +239,7 @@ fn serding() {
         use traits::TryInto;
         use starknet::ContractAddress;
         use starknet::Felt252TryIntoContractAddress;
+        use cheatcodes::{ declare, PreparedContract, deploy };
         
         #[derive(Drop, Serde)]
         struct NestedStruct {
@@ -323,6 +324,7 @@ fn proxy_storage() {
         use traits::TryInto;
         use starknet::ContractAddress;
         use starknet::Felt252TryIntoContractAddress;
+        use cheatcodes::{ declare, PreparedContract, deploy };
         
         #[derive(Drop, Serde, PartialEq, Copy)]
         struct NestedStruct {
