+++ crates/forge/tests/integration/mod.rs
@@ -3,6 +3,7 @@ mod declare;
 mod deploy;
 mod dispatchers;
 mod prank;
+mod pure_cairo;
 mod roll;
 mod starknet;
 mod syscalls;
