+++ crates/forge/tests/integration/pure_cairo.rs
@@ -1,0 +1,53 @@
+use crate::integration::common::corelib::{corelib_path, predeployed_contracts};
+use crate::{assert_failed, assert_passed, test_case};
+use camino::Utf8PathBuf;
+use forge::run;
+use indoc::indoc;
+
+#[test]
+fn simple() {
+    let test = test_case!(indoc!(
+        r#"#[test]
+        fn test_two_and_two() {
+            assert(2 == 2, '2 == 2');
+        }
+    "#
+    ));
+
+    let result = run(
+        &test.path().unwrap(),
+        &test.path().unwrap().join("src/lib.cairo"),
+        &Some(test.linked_libraries()),
+        &Default::default(),
+        &corelib_path(),
+        &test.contracts(&corelib_path()).unwrap(),
+        &Utf8PathBuf::from_path_buf(predeployed_contracts().to_path_buf()).unwrap(),
+    )
+    .unwrap();
+
+    assert_passed!(result);
+}
+
+#[test]
+fn failing() {
+    let test = test_case!(indoc!(
+        r#"#[test]
+        fn test_two_and_three() {
+            assert(2 == 3, '2 == 3');
+        }
+    "#
+    ));
+
+    let result = run(
+        &test.path().unwrap(),
+        &test.path().unwrap().join("src/lib.cairo"),
+        &Some(test.linked_libraries()),
+        &Default::default(),
+        &corelib_path(),
+        &test.contracts(&corelib_path()).unwrap(),
+        &Utf8PathBuf::from_path_buf(predeployed_contracts().to_path_buf()).unwrap(),
+    )
+    .unwrap();
+
+    assert_failed!(result);
+}
