+++ crates/forge/tests/integration/roll.rs
@@ -17,8 +17,7 @@ fn start_roll_simple() {
             use traits::TryInto;
             use starknet::ContractAddress;
             use starknet::Felt252TryIntoContractAddress;
-            use cheatcodes::PreparedContract;
-            use forge_print::PrintTrait;
+            use cheatcodes::{ declare, PreparedContract, deploy, start_roll };
             
             #[starknet::interface]
             trait IRollChecker<TContractState> {
@@ -71,7 +70,7 @@ fn start_roll_with_other_syscall() {
             use traits::TryInto;
             use starknet::ContractAddress;
             use starknet::Felt252TryIntoContractAddress;
-            use cheatcodes::PreparedContract;
+            use cheatcodes::{ declare, PreparedContract, deploy, start_roll };
             
             #[starknet::interface]
             trait IRollChecker<TContractState> {
@@ -126,8 +125,7 @@ fn start_roll_in_constructor_test() {
             use traits::TryInto;
             use starknet::ContractAddress;
             use starknet::Felt252TryIntoContractAddress;
-            use cheatcodes::PreparedContract;
-            use forge_print::PrintTrait;
+            use cheatcodes::{ declare, PreparedContract, deploy, start_roll };
             
             #[starknet::interface]
             trait IConstructorRollChecker<TContractState> {
@@ -140,7 +138,7 @@ fn start_roll_in_constructor_test() {
                 let prepared = PreparedContract { class_hash: class_hash, constructor_calldata: @ArrayTrait::new() };
                 let contract_address: ContractAddress = 2598896470772924212281968896271340780432065735045468431712403008297614014532.try_into().unwrap();
                 start_roll(contract_address, 234);
-                let contract_address: ContractAddress = deploy(prepared).unwrap().try_into().unwrap();
+                let contract_address = deploy(prepared).unwrap();
             
                 let dispatcher = IConstructorRollCheckerDispatcher { contract_address };
                 assert(dispatcher.get_stored_block_number() == 234, 'Wrong stored blk_nb');
@@ -179,8 +177,7 @@ fn stop_roll() {
             use traits::TryInto;
             use starknet::ContractAddress;
             use starknet::Felt252TryIntoContractAddress;
-            use cheatcodes::PreparedContract;
-            use forge_print::PrintTrait;
+            use cheatcodes::{ declare, PreparedContract, deploy, start_roll, stop_roll };
             
             #[starknet::interface]
             trait IRollChecker<TContractState> {
@@ -240,8 +237,7 @@ fn double_roll() {
             use traits::TryInto;
             use starknet::ContractAddress;
             use starknet::Felt252TryIntoContractAddress;
-            use cheatcodes::PreparedContract;
-            use forge_print::PrintTrait;
+            use cheatcodes::{ declare, PreparedContract, deploy, start_roll, stop_roll };
             
             #[starknet::interface]
             trait IRollChecker<TContractState> {
@@ -303,7 +299,8 @@ fn start_roll_with_proxy() {
             use traits::Into;
             use starknet::ContractAddress;
             use starknet::Felt252TryIntoContractAddress;
-            use cheatcodes::PreparedContract;
+            use cheatcodes::{ declare, PreparedContract, deploy, start_roll };
+            
             #[starknet::interface]
             trait IRollCheckerProxy<TContractState> {
                 fn get_roll_checkers_block_info(ref self: TContractState, address: ContractAddress) -> u64;
