+++ crates/forge/tests/integration/starknet.rs
@@ -15,7 +15,7 @@ fn timestamp_doesnt_decrease_between_transactions() {
             use traits::TryInto;
             use starknet::ContractAddress;
             use starknet::Felt252TryIntoContractAddress;
-            use cheatcodes::PreparedContract;
+            use cheatcodes::{ declare, PreparedContract, deploy };
 
             #[starknet::interface]
                 trait ITimestamper<TContractState> {
@@ -101,7 +101,7 @@ fn block_doesnt_decrease_between_transactions() {
             use traits::TryInto;
             use starknet::ContractAddress;
             use starknet::Felt252TryIntoContractAddress;
-            use cheatcodes::PreparedContract;
+            use cheatcodes::{ declare, PreparedContract, deploy };
 
             #[starknet::interface]
             trait IBlocker<TContractState> {
@@ -214,7 +214,7 @@ fn nonce_increases_between_transactions() {
             use traits::TryInto;
             use starknet::ContractAddress;
             use starknet::Felt252TryIntoContractAddress;
-            use cheatcodes::PreparedContract;
+            use cheatcodes::{ declare, PreparedContract, deploy };
 
             #[starknet::interface]
             trait INoncer<TContractState> {
@@ -306,7 +306,7 @@ fn nonce_increases_between_deploys_and_declares() {
             use traits::TryInto;
             use starknet::ContractAddress;
             use starknet::Felt252TryIntoContractAddress;
-            use cheatcodes::PreparedContract;
+            use cheatcodes::{ declare, PreparedContract, deploy };
 
             #[starknet::interface]
             trait INoncer<TContractState> {
