+++ src/lib.cairo
@@ -2,7 +2,17 @@ mod cheatcodes;
 
 use cheatcodes::declare;
 use cheatcodes::deploy;
+
 use cheatcodes::PreparedContract;
+use cheatcodes::RevertedTransaction;
+use cheatcodes::RevertedTransactionTrait;
+
+use cheatcodes::start_prank;
+use cheatcodes::stop_prank;
+use cheatcodes::start_roll;
+use cheatcodes::stop_roll;
+use cheatcodes::start_warp;
+use cheatcodes::stop_warp;
 
 mod forge_print;
